Simplify rule interface by replacing state with previous tx
The only real practical reason for looking at stateful http
transactions is to look at a previous request for the same resource.

This implies that in the future we need to properly invalidate
and maintain that resource, which is something we'll probably
have the semantic awareness to do.

By removing the state from the interface, we make rules simpler.